requirements.txt
# For notebook tqdm compatibility
jupyter
ipywidgets
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
click==8.3.1
comm==0.2.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
debugpy==1.8.19
decorator==5.2.1
einops==0.8.2
executing==2.2.1
filelock==3.20.3
fla-core==0.4.1
flash-linear-attention==0.4.1
flash_attn @ file:///home/m_tes/groundthink/gt-v6/flash-attn-unpacked/kaggle/working/flash-attention/dist/flash_attn-2.8.3-cp312-cp312-linux_x86_64.whl#sha256=68ad03527ff1073019a891c196262b36cdc1939e5a924ca422c579b08838c9c9
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
safetensors==0.7.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
triton==3.6.0
typer-slim==0.21.1
typing_extensions==4.15.0
wcwidth==0.5.0
wheel==0.46.3
